pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inkotter"
version = "0.1.0"
description = "Open printing for closed devices."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.0",
  "CairoSVG>=2.7",
]

[project.scripts]
inkotter = "inkotter.app.cli:main"
inkotter-gui = "inkotter.app.gui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
